requirements.txt
# Database
pymongo>=4.3.3
psycopg2-binary>=2.9.5
redis>=4.5.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.2

# Utilities
tqdm>=4.65.0